discovery-ui-service/src/main/java/com/jubl/food/nextgendiscovery/service/config/MvcConfig.java
package com.jubl.food.nextgendiscovery.service.config;

import org.springframework.context.MessageSource;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.context.support.ResourceBundleMessageSource;
import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
import org.springframework.web.servlet.i18n.LocaleChangeInterceptor;

/**
 * Web mvc configuration.
 *
 * @author abhishek
 * @version 1.0.0
 */
@Configuration
public class MvcConfig implements WebMvcConfigurer {

    /**
     * message Source.
     * @return MessageSource
     */
    @Bean
    public MessageSource messageSource() {
        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
        messageSource.setBasename("i18n/langdata");
        messageSource.setDefaultEncoding("UTF-8");
        return messageSource;
    }

    /**
     * locale Change Interceptor.
     * @return LocaleChangeInterceptor
     */
    @Bean
    public LocaleChangeInterceptor localeChangeInterceptor() {
        LocaleChangeInterceptor lci = new LocaleChangeInterceptor();
        lci.setParamName("lang");
        return lci;
    }

    /**
     * add Interceptors method.
     */
    @Override
    public void addInterceptors(final InterceptorRegistry registry) {
        registry.addInterceptor(localeChangeInterceptor());
    }

}
